Move app out and set up folder aliases
