+++ bmh_lims/frontend/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Route, Switch } from 'react-router-dom'
-import { GlobalStyle } from '../globalStyles'
+import { GlobalStyle } from './globalStyles'
 
-import { HomePage } from '../pages'
+import { HomePage } from 'pages'
 
 const App = () => {
     return (
