+++ bmh_lims/frontend/src/assets/index.js
@@ -1,4 +1,4 @@
-export { default as Logo } from './logo.svg'
-export { default as SampleIcon } from './samples.svg'
-export { default as StatusIcon } from './status.svg'
-export { default as WorkflowsIcon} from './workflows.svg'
+export { default as Logo } from './icons/logo.svg'
+export { default as SampleIcon } from './icons/samples.svg'
+export { default as StatusIcon } from './icons/status.svg'
+export { default as WorkflowsIcon} from './icons/workflows.svg'
