+++ bmh_lims/frontend/src/components/index.js
@@ -1,2 +1,2 @@
 export { default as Button } from './Button'
-export { IconButton } from './IconButton'
+export { default as IconButton } from './IconButton'
