+++ bmh_lims/frontend/src/globalStyles.js
@@ -1,5 +1,10 @@
 import { createGlobalStyle } from 'styled-components'
 
+/* 
+Styles for the entire app. This is injected in the App component.
+
+*/
+
 export const GlobalStyle = createGlobalStyle`
     body {
         font-family: 'Assistant', sans-serif;
