+++ bmh_lims/frontend/src/index.js
@@ -1,4 +1,4 @@
-import App from './components/App'
+import App from './App'
 import React from 'react'
 import { BrowserRouter} from 'react-router-dom'
 import ReactDOM from 'react-dom'
